refactor

+++ src/main/kotlin/ru/quipy/payments/logic/PaymentExternalServiceImpl.kt
@@ -37,7 +37,7 @@ class PaymentExternalSystemAdapterImpl(
     private val client = OkHttpClient.Builder().build()
 
     private val rateLimiter = SlidingWindowRateLimiter(rate = rateLimitPerSec.toLong(), window = Duration.ofSeconds(1))
-    private val semaphore = Semaphore((parallelRequests).toInt())
+    private val semaphore = Semaphore((parallelRequests))
 
     override fun performPaymentAsync(paymentId: UUID, amount: Int, paymentStartedAt: Long, deadline: Long) {
         logger.warn("[$accountName] Submitting payment request for payment $paymentId")
